+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Changed
 * Update README with gem release instructions
+* First implementation of YasstFile encrypt/decrypt actions
 
 ## [0.1.1] - 2016-02-11
 ### Changed